+++ packages/tools/src/index.ts
@@ -22,6 +22,7 @@ export {
   handleUploadResponse,
 } from './lib/build-cache/fetchCachedBuild.js';
 export { getInfoPlist } from './lib/getInfoPlist.js';
+export { getReactNativeVersion } from './lib/getReactNativeVersion.js';
 export * from './lib/fs.js';
 export * from './lib/package-json.js';
 export * from './lib/project-name.js';